requirements.txt
llama-index==0.12.22
llama-index-llms-openrouter==0.3.0
llama-index-embeddings-huggingface==0.4.0
llama-index-readers-web==0.3.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
pypdf==5.3.1
sentence-transformers==3.1.1
pandas==2.2.3